yoke: make turbulence less senstive to metadata properties
