+++ internal/revision.go
@@ -7,7 +7,6 @@ import (
 	"fmt"
 	"net/url"
 	"path"
-	"reflect"
 	"slices"
 	"strconv"
 	"strings"
@@ -123,36 +122,6 @@ func GetOwner(resource *unstructured.Unstructured) string {
 	}
 
 	return namespace + "/" + release
-}
-
-// RemoveAdditions removes fields from actual that are not in expected.
-// it removes the additional properties in place and returns "actual" back.
-// Values passed to removeAdditions are expected to be generic json compliant structures:
-// map[string]any, []any, or scalars.
-func RemoveAdditions[T any](expected, actual T) T {
-	// Check if we can access the types safely
-	if !reflect.ValueOf(expected).IsValid() || !reflect.ValueOf(actual).IsValid() || reflect.ValueOf(actual).Type() != reflect.ValueOf(expected).Type() {
-		return actual
-	}
-
-	switch a := any(actual).(type) {
-	case map[string]any:
-		e := any(expected).(map[string]any)
-		for key := range a {
-			if _, ok := e[key]; !ok {
-				delete(a, key)
-				continue
-			}
-			a[key] = RemoveAdditions(e[key], a[key])
-		}
-	case []any:
-		e := any(expected).([]any)
-		for i := range min(len(a), len(e)) {
-			a[i] = RemoveAdditions(e[i], a[i])
-		}
-	}
-
-	return actual
 }
 
 func Canonical(resource *unstructured.Unstructured) string {
